extraction/requirements.txt
fastapi
uvicorn
pydantic
lxml
networkx
pymupdf
numpy
google-generativeai
easyocr
docling[vlm]
transformers
accelerate
requests
beautifulsoup4
python-dotenv
reportlab
